Push_swap initial commit
Implemented a program to generate random numbers.

Created a program called checker that get a environment variable with a
list of numbers as a parameter. It should create a stack with this
numbers in the same order in which they were passed, and other empty
stack. Then it should read some instructions from standard imput and
execute them. After that if the first stack has the numbers well ordered shoud print OK on the standard output. If not, should print KO. Also should check for errors.

Created all the instructions that can be used in the program. And the
tools to manage the linked lists used as memory stack.