More list management functions and first algorithm
This first algorithm get the average value and sends to the stack b the
elements that are smaller (usually around half of them). Then starts
again geting the average from the other half and sending the smaller
elements to stack b. It continues doing this untill there is no more
elements. The elements bigger than average value are sent to the end of
the list. Each time an element is sent to the stack b it checks if the
value is smaller than the second one in the stack. In that case it swaps
that elements.

Once all elements are in stack b, it does the same but starting from the
last one pushed to stack b. If the element is bigger than the average is
sent back to the stack a. If not, it is sent to the end of the stack b.

This algorithm should be repited until the stack a is sorted.

It works, but it takes too much steps (more than 3000 for 100 numbers).